# place2connect-api/docs/swagger.yaml
definitions:
...
  handlers.CommentRequest:
    properties:
      commentDescription:
        minLength: 5
        type: string
      postID:
        type: integer
      userID:
        type: integer
    required:
    - commentDescription
    - postID
    - userID
    type: object
  handlers.CommentResponse:
    properties:
      comment_description:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  handlers.LikeRequest:
    properties:
      isLikablePost:
        type: boolean
      postID:
        type: integer
      userID:
        type: integer
    required:
    - isLikablePost
    - postID
    - userID
    type: object
  handlers.LikeResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_likable_post:
        type: boolean
      post_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  handlers.PostListResponse:
    properties:
      postCount:
        type: integer
      posts:
        items:
          $ref: '#/definitions/handlers.PostResponse'
        type: array
    type: object
  handlers.PostRequest:
    properties:
      description:
        description: |-
          FirstName       string `json:"first_name" validate:"required,min=3"`
          LastName        string `json:"last_name" validate:"required,min=3"`
          Location        string `json:"location,omitempty"`
        minLength: 8
        type: string
      picturePath:
        type: string
      userID:
        type: integer
    required:
    - description
    - userID
    type: object
  handlers.PostResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/handlers.CommentResponse'
        type: array
      createdAt:
        type: string
      description:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      likes:
        items:
          $ref: '#/definitions/handlers.LikeResponse'
        type: array
      location:
        type: string
      picturePath:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
      userPicturePath:
        type: string
    type: object
...
info:
  contact:
    email: contact@owonafx.com
    name: PLACE2CONNECT API Support
  description: PLACE2CONNECT Docs.
  license:
    name: pace2connect 2.0
  termsOfService: http://swagger.io/terms/
  title: PLACE2CONNECT API
  version: "1.0"
paths:
...
  /api/v1/imageUpload/:
    post:
      consumes:
      - application/json
      description: UploadImage func to upload an image
      operationId: upload-image
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Upload an image
      tags:
      - Post
  /api/v1/posts:
    post:
      consumes:
      - application/json
      description: CreatePost create a new Post
      operationId: create-post
      parameters:
      - description: The PostRequest
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/handlers.PostRequest'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PostResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create a new Post
      tags:
      - Post
  /api/v1/posts/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Display all Post. Auth is required
      operationId: all-posts
      parameters:
      - description: Limit number of posts returned (default is 20)
        in: query
        name: limit
        type: integer
      - description: Offset/Skip number of posts (default is 0)
        in: query
        name: offset
        type: integer
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PostListResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "500":
          description: Status Internal Server Error
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Display all Post
      tags:
      - Post
  /api/v1/posts/users/{user_id}/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Display all Posts of a User By userId. Auth is required
      operationId: all-user-posts
      parameters:
      - description: ID of the user for posts
        in: path
        name: user_id
        required: true
        type: integer
      - description: Limit number of Posts returned (default is 20)
        in: query
        name: limit
        type: integer
      - description: Offset/Skip number of posts (default is 0)
        in: query
        name: offset
        type: integer
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PostListResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "500":
          description: Status Internal Server Error
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Display all Posts of a User
      tags:
      - User
  /api/v1/usercomments:
    post:
      consumes:
      - application/json
      description: AddCommentToPost add a new comment to Post
      operationId: add-comment
      parameters:
      - description: CommentRequest
        in: body
        name: commentRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.CommentRequest'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PostResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add a new Comment to a Post
      tags:
      - Post
...
  /api/v1/userlikes:
    post:
      consumes:
      - application/json
      description: AddLikeToPost add a new like to Post
      operationId: add-like
      parameters:
      - description: LikeRequest
        in: body
        name: likeRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.LikeRequest'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PostResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add a new Like to a Post
      tags:
      - Post
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"